Create LICENSE
This repository will be archived. Release previously unlicensed, original code under MIT. Certain files generated by GNU AutoTools (that really should not be in this repo to begin with) are released under GPL-2. Those files maintain their original copyright/license text in the file header. The GPL-2 License text to be included in a separate commit.